fix: repair broken botorch imports and address PR review findings
Critical:
- Fix stale `alf_tools.utils.botorch_utils` import in BoTorchGPModel that
  broke collection of the entire tools test suite after the module was
  relocated to alf_tools.models.utils during a main merge.

Model:
- Wire up normalize_inputs/standardize_outputs via BoTorch Normalize and
  Standardize transforms (previously stored but never applied) and remove
  the misleading "automatic standardization" comment.
- Default dtype to torch.float64 for GP numerical stability; update tests.

Cleanup:
- Delete dead duplicate modules: BoTorchModelAdapter (superseded by
  BotorchModelWrapper) and ContinuousSearch (superseded by
  BotorchContinuousSearch), plus the adapter test.
- Add multi-objective dataset tests (branincurrin, dtlz2, zdt1).
- Remove duplicated Apache license headers, fix docstring import paths,
  collapse a redundant branch, annotate type: ignore, resolve stale TODOs.

+++ tools/tests/models/test_botorch_exact_gp_model.py
@@ -12,20 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# Copyright 2023 InstaDeep Ltd. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 """Tests for BoTorchGPModel class.
 
 This module tests the BoTorch-based GP model implementation, including:
@@ -97,7 +83,7 @@ class TestBoTorchGPModelInitialization:
         assert model.learning_rate == 0.1
         assert model.optimizer == "scipy"
         assert model.max_attempts == 5
-        assert model.dtype == torch.float32
+        assert model.dtype == torch.float64
         assert model.model is None
         assert model.train_X is None
         assert model.train_Y is None
@@ -160,7 +146,7 @@ class TestBoTorchGPModelFeaturisation:
 
         assert isinstance(features, torch.Tensor)
         assert features.shape == (3, 2)
-        assert features.dtype == torch.float32
+        assert features.dtype == torch.float64
 
     def test_featurise_single_candidate(self):
         """Test featurisation with single candidate."""
